Refactor MVP
