+++ app/src/main/java/com/myapps/dogsrecycler/ui/fragments/AnimalFragment.kt
@@ -9,8 +9,8 @@ import android.view.ViewGroup
 import com.arellomobile.mvp.MvpAppCompatFragment
 import com.arellomobile.mvp.presenter.InjectPresenter
 import com.myapps.dogsrecycler.R
-import com.myapps.dogsrecycler.common.InfiniteScrollListener
-import com.myapps.dogsrecycler.common.inflate
+import com.myapps.dogsrecycler.ui.common.InfiniteScrollListener
+import com.myapps.dogsrecycler.utils.inflate
 import com.myapps.dogsrecycler.ui.adapters.AnimalAdapter
 import kotlinx.android.synthetic.main.fragment_animals.*
 
@@ -40,7 +40,12 @@ class AnimalFragment : MvpAppCompatFragment(),
             setHasFixedSize(true)
             layoutManager = linearLayout
             clearOnScrollListeners()
-            addOnScrollListener(InfiniteScrollListener({ requestAnimals() }, linearLayout))
+            addOnScrollListener(
+                InfiniteScrollListener(
+                    { requestAnimals() },
+                    linearLayout
+                )
+            )
         }
         animalPresenter.animal = animal
